Only restaurant owner can modify or update their restaurant

+++ src/controllers/RestaurantController.ts
@@ -98,19 +98,23 @@ export const getOne: Handler = async (req, res) => {
 export const modify: Handler = async (req, res) => {
   if (!req.user) return res.status(400).send('User not authenticated')
   if (req.user.role === 'restaurateur') {
-    req.body.owner = req.user
-    try {
-      const Restaurant = await RestaurantModel.findOneAndUpdate(
-        { _id: req.params.id },
-        { $set: req.body },
-        { new: true }
-      )
-      if (Restaurant) res.send(Restaurant)
-      else res.status(404).send('Restaurant Not Found')
-    } catch (err) {
-      if (err instanceof Error && err.message) res.status(400).send(err.message)
-      else throw err
-    }
+    const ownerRestaurant = await RestaurantModel.find({ 'owner._id': req.user._id }, { projection: { _id: 1 } })
+    if (ownerRestaurant.length === 0) return res.status(400).send('Restaurant Not Found')
+    if (ownerRestaurant[0]._id === req.params.id) {
+      req.body.owner = req.user
+      try {
+        const Restaurant = await RestaurantModel.findOneAndUpdate(
+          { _id: req.params.id },
+          { $set: req.body },
+          { new: true }
+        )
+        if (Restaurant) res.send(Restaurant)
+        else res.status(404).send('Restaurant Not Found')
+      } catch (err) {
+        if (err instanceof Error && err.message) res.status(400).send(err.message)
+        else throw err
+      }
+    } else return res.status(400).send('You are not the owner of this restaurant')
   } else {
     return res.status(400).send('User is not a restaurateur, please login to a restaurateur account')
   }
@@ -119,9 +123,14 @@ export const modify: Handler = async (req, res) => {
 export const remove: Handler = async (req, res) => {
   if (!req.user) return res.status(400).send('User not authenticated')
   if (req.user.role === 'restaurateur') {
-    const Restaurant = await RestaurantModel.deleteOne({ _id: req.params.id })
-    if (Restaurant.deletedCount) res.sendStatus(204)
-    else res.status(404).send('Restaurant Not Found')
+    const ownerRestaurant = await RestaurantModel.find({ 'owner._id': req.user._id }, { projection: { _id: 1 } })
+    if (ownerRestaurant.length === 0) return res.status(400).send('Restaurant Not Found')
+    if (ownerRestaurant[0]._id === req.params.id) {
+      req.body.owner = req.user
+      const Restaurant = await RestaurantModel.deleteOne({ _id: req.params.id })
+      if (Restaurant.deletedCount) res.sendStatus(204)
+      else res.status(404).send('Restaurant Not Found')
+    }
   } else {
     return res.status(400).send('User is not a restaurateur, please login to a restaurateur account')
   }
